+++ pycm/pycm_overall_func.py
@@ -9,6 +9,107 @@ from .pycm_ci import kappa_SE_calc, CI_calc, SE_calc
 from .pycm_util import complement
 
 
+def alpha2_calc(TOP, P, ACC, POP, classes, max_iter=200, epsilon=0.0001):
+    """
+    Calculate Aickin's alpha.
+
+    :param TOP: test outcome positive
+    :type TOP : dict
+    :param P: condition positive
+    :type P : dict
+    :param ACC: accuracy
+    :type ACC : float
+    :param POP: population
+    :type POP : dict
+    :param classes: confusion matrix classes
+    :type classes : list
+    :param max_iter: maximum iteration
+    :type max_iter: int
+    :param epsilon: difference threshold
+    :type epsilon: float
+    :return: Aickin's alpha as float
+    """
+    try:
+        p_A = {i: TOP[i] / POP[i] for i in classes}
+        p_B = {i: P[i] / POP[i] for i in classes}
+        step = 1
+        alpha = 0
+        alpha_prev = 0
+        while(True):
+            p_e = 0
+            for i in classes:
+                p_e += (p_A[i] * p_B[i])
+            alpha_prev = alpha
+            alpha = reliability_calc(p_e, ACC)
+            for i in classes:
+                p_A[i] = TOP[i] / \
+                    (((1 - alpha) + alpha * p_B[i] / p_e) * POP[i])
+                p_B[i] = P[i] / (((1 - alpha) + alpha * p_A[i] / p_e) * POP[i])
+            if step > max_iter or abs(alpha - alpha_prev) < epsilon:
+                break
+            step += 1
+        return alpha
+    except Exception:
+        return "None"
+
+
+def alpha_calc(RACC, ACC, POP):
+    """
+    Calculate unweighted Krippendorff's alpha.
+
+    :param RACC: random accuracy
+    :type RACC : float
+    :param ACC: accuracy
+    :type ACC : float
+    :param POP: population
+    :type POP : int
+    :return: unweighted alpha as float
+    """
+    try:
+        epsi = 1 / (2 * POP)
+        p_a = (1 - epsi) * ACC + epsi
+        p_e = RACC
+        return reliability_calc(p_e, p_a)
+    except Exception:
+        return "None"
+
+
+def weighted_alpha_calc(classes, table, P, TOP, POP, weight):
+    """
+    Calculate weighted Krippendorff's alpha.
+
+    :param classes: confusion matrix classes
+    :type classes : list
+    :param table: confusion matrix table
+    :type table : dict
+    :param P: condition positive
+    :type P : dict
+    :param TOP: test outcome positive
+    :type TOP : dict
+    :param POP: population
+    :type POP : dict
+    :param weight: weight matrix
+    :type weight: dict
+    :return: weighted alpha as float
+    """
+    p_e = 0
+    p_a = 0
+    population = list(POP.values())[0]
+    epsi = 1 / (2 * population)
+    try:
+        w_max = max(map(lambda x: max(x.values()), weight.values()))
+        for i in classes:
+            for j in classes:
+                v_i_j = 1 - weight[i][j] / w_max
+                p_e += (((P[i] + TOP[j]) / (POP[i] * 2)) ** 2) * v_i_j
+                p_a += table[i][j] * v_i_j / POP[i]
+        p_a = (1 - epsi) * p_a + epsi
+        weighted_alpha = reliability_calc(p_e, p_a)
+        return weighted_alpha
+    except Exception:
+        return "None"
+
+
 def B_calc(classes, TP, TOP, P):
     """
     Calculate B (Bangdiwala's B).
@@ -391,15 +492,15 @@ def weighted_kappa_calc(classes, table, P, TOP, POP, weight):
     :return: weighted kappa as float
     """
     p_e = 0
-    p_s = 0
+    p_a = 0
     try:
         w_max = max(map(lambda x: max(x.values()), weight.values()))
         for i in classes:
             for j in classes:
                 v_i_j = 1 - weight[i][j] / w_max
                 p_e += P[i] * TOP[j] * v_i_j / (POP[i]**2)
-                p_s += table[i][j] * v_i_j / POP[i]
-        weighted_kappa = reliability_calc(p_e, p_s)
+                p_a += table[i][j] * v_i_j / POP[i]
+        weighted_kappa = reliability_calc(p_e, p_a)
         return weighted_kappa
     except Exception:
         return "None"
@@ -875,6 +976,10 @@ def overall_statistics(**kwargs):
     TNR_micro = micro_calc(item1=kwargs["TN"], item2=kwargs["FP"])
     TNR_macro = macro_calc(kwargs["TNR"])
     B = B_calc(classes, TP, TOP, P)
+    alpha = alpha_calc(
+        overall_random_accuracy_unbiased,
+        overall_accuracy,
+        population)
     return {
         "Overall ACC": overall_accuracy,
         "Kappa": overall_kappa,
@@ -937,4 +1042,5 @@ def overall_statistics(**kwargs):
         "Pearson C": C,
         "CSI": CSI,
         "ARI": ARI,
-        "Bangdiwala B": B}
+        "Bangdiwala B": B,
+        "Krippendorff Alpha": alpha}